Changes in Impala-lzo for the compressed text reader
The compressed text reader patch, changes a bit the variables of the HdfsScanner and
HdfsTextScanner. Also, it removes the LZO_TEXT file format (it was not a file format,
rather a convention to identify lzo-compressed text files), and moves the INDEX_SUFFIX to
HdfsTextScanner. Finally, it changes the name of the IssueInitialRanges wrapper.
These changes are needed for Impala-lzo to compile.

Change-Id: Ibd6f5e0b16465806ef4be24e1451a2817f5f72d4
(cherry picked from commit 18f79623dd41ed36a8d7922431f3aba142e146ff)