Fix HdfsLzoTextScanner::Close()
The LZO scanner did not call HdfsTextScanner::Close(), instead
replicating the logic in its own Close() method. This makes it harder
to keep in sync with the text scanner, such as in this case, where it
was not closing the new boundary_pool_. This patch has the LZO scanner
call HdfsTextScanner::Close() instead of replicating it.

Change-Id: I65c2117984922b4eb3ca25e44433dbfffbd2874c