IMPALA-3832: fix DCHECK calling AddDetail() on an OK status
The bug is that it's invalid to call AddDetail() without first
initialising the status to an error status.

Instead of trying to aggregate all errors in the LZO text file
header, simplify control flow by returning as soon as we find
that the file is invalid.

Testing:
Will enable fuzz testing for the lzo text scanner before
closing the JIRA. https://gerrit.cloudera.org/#/c/3448/

Change-Id: Idc3887e4079ff8df139a91df71ca695bdc0736f4