IMPALA-4262: LZO-scanner fails when reading large index files from S3
The LZO-scanner always expects the index file to be a multiple of
sizeof(uint64_t). In a loop, it reads the file 10MB at a time
(which is a multiple of sizeof(uint64_t)) and processes this buffer
uint64_t bytes at a time. This worked fine so far for HDFS as HDFS
returns always returned the entire buffer as requested. However, this
is not a valid assumption.

For S3, the internal read() relies on a stream, therefore it
sends as many bytes as currently available. This means that the
'bytes read' may not be a multiple of sizeof(uint64_t). This currently
causes the scanner to throw an error in this case.

This patch modifies the LZO-scanner logic so that it carries over
(buffer_size % sizeof(uint64_t)) bytes every time and processes these
remaining bytes as a part of the next iteration of the read loop.

Change-Id: I9874d0d5dbcf64e1afeb9cfe655f7b9c58c640ed