IMPALA-4322: test_scanners_fuzz.py hits a DCHECK
During fuzz testing, we discovered an edge case that hit a DCHECK.
This case should never show up in the real world unless working with
a malformed LZO index file.

When the file is less than 8 bytes in size, the DCHECK would be hit as
it previously assumed that would never happen. The DCHECK is now removed.

Change-Id: Iff5145c7f79dc6003c42a37dad4125517831410c